+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ This project adheres to `Semantic Versioning <http://semver.org/>`__.
 [Unreleased]
 ------------
 
+- Removed kicked-user-made-bot-think-he-was-kicked bug
+
 [1.0] - 2016-04-14
 ------------------